Allow css transformations of title in config
By default the style uppercases the title, which forces a
style choice to the user.

We can add extra css but that might be overkill for just one
change: For example, not capitalizing the title default.

This allows the title to be fully lowercase, by simply overriding
the theme configuration.